+++ .eslintrc.js
@@ -1,0 +1,56 @@
+const { peerDependencies } = require('./package.json');
+
+module.exports = {
+  parser: '@typescript-eslint/parser',
+  parserOptions: {
+    createDefaultProgram: true,
+    project: './tsconfig.json',
+  },
+  plugins: ['@typescript-eslint', 'react-hooks'],
+  extends: [
+    'airbnb',
+    'plugin:@typescript-eslint/recommended',
+    'plugin:@typescript-eslint/recommended-requiring-type-checking',
+    'prettier',
+    'prettier/@typescript-eslint',
+  ],
+  settings: {
+    'import/resolver': {
+      node: {
+        extensions: ['.ts', '.tsx'],
+      },
+    },
+  },
+  env: {
+    browser: true,
+  },
+  rules: {
+    '@typescript-eslint/array-type': 'off',
+    '@typescript-eslint/explicit-member-accessibility': ['warn', { accessibility: 'no-public' }],
+    '@typescript-eslint/no-empty-function': 'error',
+    '@typescript-eslint/no-empty-interface': 'off',
+    '@typescript-eslint/no-explicit-any': 'off',
+    '@typescript-eslint/no-parameter-properties': 'off',
+    '@typescript-eslint/no-use-before-define': 'off',
+    '@typescript-eslint/no-useless-constructor': 'error',
+    '@typescript-eslint/unbound-method': 'off',
+    'import/extensions': 'off',
+    'import/no-unresolved': ['error', { ignore: Object.keys(peerDependencies) }],
+    'import/prefer-default-export': 'off',
+    'jsx-a11y/accessible-emoji': 'off',
+    'jsx-a11y/anchor-is-valid': 'off',
+    'jsx-a11y/label-has-associated-control': 'off',
+    'no-continue': 'off',
+    'no-empty-function': 'off',
+    'no-plusplus': 'off',
+    'no-useless-constructor': 'off',
+    'react-hooks/exhaustive-deps': 'warn',
+    'react-hooks/rules-of-hooks': 'error',
+    'react/jsx-curly-brace-presence': 'off',
+    'react/jsx-filename-extension': ['warn', { extensions: ['.tsx'] }],
+    'react/jsx-one-expression-per-line': 'off',
+    'react/jsx-props-no-spreading': 'off',
+    'react/jsx-wrap-multilines': 'off',
+    'react/prop-types': 'off',
+  },
+};